Refactor assignment button into its own component
